requirements.txt
fastapi
uvicorn
aiogram
openai
sqlalchemy
asyncpg
python-dotenv
httpx
greenlet